pyproject.toml
[tool.poetry]
name = "docstring-ai"
version = "0.0.1"
description = "Add docstrings to Python files in a repository using OpenAI Assistants API and ChromaDB with caching."
authors = ["Pierre-Henri AUsSEIL <ph.ausseil@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ph-ausseil/docstring-ai"
repository = "https://github.com/ph-ausseil/docstring-ai"
keywords = ["docstrings", "GitHub", "automation"]

[tool.poetry.scripts]
docstring-ai = "docstring_ai.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.0.0"           
chromadb = "^0.5.0"         
tiktoken = "<1.0.0"          
python-dotenv = "*"     
PyGithub = "^1.0.0"
onnxruntime = "<1.20.0"
numpy="<2.0.0"

[tool.poetry.dev-dependencies]
black = "*"
flake8 = "*"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
